requirements.txt
fastapi
uvicorn
python-telegram-bot==21.0
openai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pypdf
python-docx
numpy
